requirements.txt
requests~=2.32.3
google-auth~=2.34.0